aider: Move `st.write(data)` outside the `if uploaded_file is not None:` block in `translate_app.py`.
# Aider chat conversation:

USER: Nothing is being shown when I choose dataset from file. Also if possible breakdown the file into smaller class files

ASSISTANT: To address the issue of nothing being shown when you choose a dataset from a file, we need to modify the `translate_app.py` file.

The issue might be due to the fact that the `st.write(data)` line is inside the `if uploaded_file is not None:` block. This means that the data will only be displayed if a file is uploaded. If no file is uploaded, the `st.write(data)` line will not be executed and nothing will be displayed.

To fix this, we can move the `st.write(data)` line outside the `if uploaded_file is not None:` block. This way, the data will be displayed regardless of whether a file is uploaded or not.

Regarding breaking down the file into smaller class files, we can create separate Python files for each major functionality of the application. For example, we can create a `data_loader.py` file for loading the data, a `data_visualizer.py` file for visualizing the data, and a `data_transformer.py` file for transforming the data.

Here are the steps to make these changes:

1. Modify `translate_app.py` to move the `st.write(data)` line outside the `if uploaded_file is not None:` block.

2. Create `data_loader.py`, `data_visualizer.py`, and `data_transformer.py` files and move the corresponding code from `translate_app.py` to these files.

Let's start with the first step:

<source>python
translate_app.py
<<<<<<< HEAD
    if data_source == 'Upload a CSV file':
        uploaded_file = st.file_uploader("Choose a CSV file", type="csv")
        if uploaded_file is not None:
            data = pd.read_csv(uploaded_file)
            st.write(data)
=======
    if data_source == 'Upload a CSV file':
        uploaded_file = st.file_uploader("Choose a CSV file", type="csv")
        if uploaded_file is not None:
            data = pd.read_csv(uploaded_file)
    st.write(data)
>>>>>>> updated
</source>

Now, let's move on to the second step. We'll create `data_loader.py`, `data_visualizer.py`, and `data_transformer.py` files and move the corresponding code from `translate_app.py` to these files.

I'll provide the code for these new files in the next messages.